Android touchable wrapper moved to Button component.
